Renaming and fixing duplicate content (#8)
* Correcting CODE_OF_CONDUCT.md name

Github assumes underscores instead of dashes, oops.  Renaming the file so
it conforms to tooling expectations.

Signed-off-by: John Hawley (VMware) <jhawley@vmware.com>

* Removing duplicate copy of the content

Copy/Paste mistake lead to there being two copies of the content
in the document, going ahead and removing the second copy.

Signed-off-by: John Hawley (VMware) <jhawley@vmware.com>

* Remove double @

One of the e-mail addresses has a double @ in it, which makes it not
copy and pasteable.  Removing one of them.

Signed-off-by: John Hawley (VMware) <jhawley@vmware.com>