Create a default security policy for this org
This change create a security policy that will effectively be the default for
all repos in the vmware org. The contents of this policy do not mention project
specific details so as to keep it generic enough to be used by any repository
in the org.

For more information on GitHub security policies, see: 
https://docs.github.com/en/code-security/getting-started/adding-a-security-policy-to-your-repository

This file serves as one of the default community health files for the org. Further details can
be found here:
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file